Makefile: don't mount ../katzenpost
because mounting it creates it if it doesn't exist, and some developers don't
keep their checkouts next to eachother. so instead, for now the procedure for
using a local katzenpost checkout in docker is to clone it inside the catchat
repo as described in the recent commits to README.md.