update and pin gio to fork with foreground permission
gio (and now catchat) require go 1.17 to build and a few api changes.

This commit resolves api changes and differences introduced by go 1.17.

This also pins to a fork of gio that uses the android foreground
permission to prevent catchat from being killed when not in foreground.

This is tracked upstream: https://github.com/gioui/gio/pull/67